Refactor geometry types
