+++ app/components/project-geometries.js
@@ -4,7 +4,6 @@ import mapboxgl from 'mapbox-gl';
 import { service } from '@ember-decorators/service';
 import { argument } from '@ember-decorators/argument';
 import { tagName } from '@ember-decorators/component';
-import { camelize } from '@ember/string';
 import { developmentSiteLayer } from './project-geometries/types/development-site';
 import { projectAreaLayer } from './project-geometries/types/project-area';
 import projectGeometryIcons from '../utils/project-geom-icons';
@@ -156,20 +155,6 @@ export default class ProjectGeometryEditComponent extends Component {
   @computed('lat', 'lng')
   get center() {
     return [this.get('lat'), this.get('lng')];
-  }
-
-  @computed('type')
-  get projectGeometryType() {
-    return `project-geometries/types/${this.get('type')}`;
-  }
-
-  @computed('type')
-  get geometricPropertyForType() {
-    const model = this.get('model');
-    const typeName = camelize(this.get('type'));
-
-    return model.get('geometricProperties')
-      .findBy('geometryType', typeName);
   }
 
   @action
