+++ app/templates/components/project-geometries.hbs
@@ -21,6 +21,7 @@
       )
       mapLoaded=(action 'handleMapLoad') as |map|}}
 
+      {{!-- Wrap drawable components and pass as context --}}
       {{#mapbox-gl-draw map=map as |drawableMap|}}
         {{search-handler
           map=map}}
@@ -54,28 +55,30 @@
         {{/map.source}}
 
         {{!-- invoke the component for the specified type --}}
-        {{component projectGeometryType
-          model=geometricPropertyForType
+        {{#project-geometries/-types
+          type=type
+          model=model
           mode=mode
-          map=drawableMap}}
-
-        {{!-- transport this markup when map is done loading --}}
-        {{#ember-wormhole to='geometry-type-save-box'}}
-          <div class="stable">
-            {{#if (eq mode 'lots')}}
-              {{#link-to 'projects.edit.geometry-edit' model.id (query-params mode='draw')
-                tagName='button'
-                class='button gray expanded'
-                disabled=(not model.hasDirtyAttributes)
-              }}
-                Switch to Manual Drawing
-                <small style="display:block;line-height:1.75;">
-                  (edit the shape with draw tools)
-                </small>
-              {{/link-to}}
-            {{/if}}
-          </div>
-        {{/ember-wormhole}}
+          map=drawableMap as |geometryType|}}
+          {{log geometryType.isReady}}
+          {{!-- transport this markup when map is done loading --}}
+          {{#ember-wormhole to='geometry-type-save-box'}}
+            <div class="stable">
+              {{#if (eq mode 'lots')}}
+                {{#link-to 'projects.edit.geometry-edit' model.id (query-params mode='draw')
+                  tagName='button'
+                  class='button gray expanded'
+                  disabled=(not geometryType.isReady)
+                }}
+                  Switch to Manual Drawing
+                  <small style="display:block;line-height:1.75;">
+                    (edit the shape with draw tools)
+                  </small>
+                {{/link-to}}
+              {{/if}}
+            </div>
+          {{/ember-wormhole}}
+        {{/project-geometries/-types}}
       {{/mapbox-gl-draw}}
     {{/labs-map}}
   {{/if}}
