+++ app/templates/components/project-geometries/-types.hbs
@@ -1,0 +1,6 @@
+{{component projectGeometryType
+  model=geometricPropertyForType
+  mode=mode
+  map=map}}
+
+{{yield (hash isReady=isReadyToProceed)}}
